Removing OS/2 - Windows type usage.  Merge my changes so far back into the trunk.  This is essentially most of the CHAR and related (PSZ, etc.)
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1110 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/instructions/EndInstruction.hpp
@@ -65,8 +65,8 @@ class RexxInstructionEnd : public RexxInstruction {
   void execute(RexxActivation *, RexxExpressionStack *);
 
   inline RexxString *endName() { return this->name; };
-  inline void        setStyle(INT type) { this->instructionInfo.flags = type; };
-  inline INT         getStyle(void)     { return this->instructionInfo.flags; };
+  inline void        setStyle(size_t type) { this->instructionInfo.flags = type; };
+  inline size_t      getStyle(void)        { return this->instructionInfo.flags; };
 
  RexxString * name;                    /* specified control variable name   */
 };
